% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shazam.R
\name{wkshzm_command}
\alias{wkshzm_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkshzm_command(shazam, cmd_json)
}
\arguments{
\item{shazam}{A shazam handle from [wkshzm_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
